+++ examples/sample-1/themes/landscape/.gitignore
@@ -1,2 +1,4 @@
 .DS_Store
 node_modules
+.yalc
+yalc.lock
