+++ examples/sample-1/themes/landscape/.npmignore
