--- filter.d.ts
@@ -1,2 +1,0 @@
-import type { Locals } from "hexo";
-export declare const register: (data: Locals.Page | Locals.Post) => Locals.Page | undefined;
