+++ index.d.ts
@@ -1,1 +1,2 @@
-export {};
+import type { Locals } from "hexo";
+export declare const register: (data: Locals.Page | Locals.Post) => Locals.Page | undefined;
