+++ package.json
@@ -4,13 +4,12 @@
   "description": "Generate breadcrumb for a Hexo page",
   "main": "index.js",
   "scripts": {
-    "build": "tsc",
+    "build": "tsc -p tsconfig.json",
     "build:watch": "tsc --watch",
     "lint": "eslint --ext .ts .",
     "format": "prettier --write **/*.{ts,md,yml,json}",
     "version": "npm run lint",
-    "postversion": "git push && git push --tags",
-    "yalc:push:watch": "watch --interval=20 'yalc push' ."
+    "postversion": "git push && git push --tags"
   },
   "repository": {
     "type": "git",
@@ -34,7 +33,7 @@
   "devDependencies": {
     "@types/hexo": "^3.8.12",
     "@typescript-eslint/eslint-plugin": "^6.14.0",
-    "eslint": "^8.55.0",
+    "eslint": "^8.56.0",
     "eslint-import-resolver-typescript": "^3.6.1",
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-prettier": "^5.0.1"
